requirements.txt
Django==4.2.1
channels==4.0.0
djangorestframework~=3.14.0
django-parler~=2.3
django-parler-rest~=2.2
requests==2.30.0
APScheduler==3.10.1
asgiref==3.6.0
greensms==2.0.1
python-dotenv==1.0.0
coreapi~=2.3.3
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==23.2
drf_yasg==1.21.5
whitenoise==6.4.0
psycopg2==2.9.6
Pillow==9.5.0
xmltodict

